Move API service provider into new API namespace.
